+++ tools/isobuild/builder.js
@@ -1,6 +1,8 @@
 import assert from "assert";
 import {WatchSet, readAndWatchFile, sha1} from '../fs/watch.js';
-import files from '../fs/files.js';
+import files, {
+  symlinkWithOverwrite,
+} from '../fs/files.js';
 import NpmDiscards from './npm-discards.js';
 import {Profile} from '../tool-env/profile.js';
 import {
@@ -393,7 +395,10 @@ Previous builder: ${previousBuilder.outputPath}, this builder: ${outputPath}`
   // A version of copyDirectory that works better for copying node_modules
   // directories when symlinks are involved.
   copyNodeModulesDirectory(options) {
-    assert.strictEqual(files.pathBasename(options.from), "node_modules");
+    // Although the options.from directory should probably be a
+    // node_modules directory, the only essential precondition here is
+    // that the destination directory is a node_modules directory.
+    // assert.strictEqual(files.pathBasename(options.from), "node_modules");
     assert.strictEqual(files.pathBasename(options.to), "node_modules");
 
     if (options.symlink) {
@@ -802,28 +807,6 @@ function symlinkIfPossible(source, target) {
     return true;
   } catch (e) {
     return false;
-  }
-}
-
-// create a symlink, overwriting the target link, file, or directory
-// if it exists
-function symlinkWithOverwrite(source, target) {
-  try {
-    files.symlink(source, target);
-  } catch (e) {
-    if (e.code === "EEXIST") {
-      // overwrite existing link, file, or directory
-      files.rm_recursive(target);
-      files.symlink(source, target);
-    } else if (e.code === "EPERM" &&
-               process.platform === "win32") {
-      files.rm_recursive(target);
-      // This will work only if source refers to a directory, but that's a
-      // chance worth taking.
-      files.symlink(source, target, "junction");
-    } else {
-      throw e;
-    }
   }
 }